feat: Funnel Step
